requirements.txt
openai>=1.0.0
elasticsearch>=8.0.0
anthropic>=0.25.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
mcp>=1.0.0
pydantic>=2.0.0